Some other EbxWriter fixes

+++ FrostySdk/IO/EbxWriter.cs
@@ -1424,7 +1424,7 @@ namespace FrostySdk.IO
                 pis.Add(pi);
             }
 
-            index = AddClass(objType.Name, fieldTypes.Count, (byte)(pis.Count + ((inherited) ? 1 : 0)),
+            index = AddClass(objType.Name, fieldTypes.Count, (ushort)(pis.Count + ((inherited) ? 1 : 0)),
                 cta.Alignment, cta.Flags, cta.Size, 0, objType);
 
             // Inherited
@@ -1982,7 +1982,7 @@ namespace FrostySdk.IO
             return (classTypes.Count - 1);
         }
 
-        private int AddClass(string name, int fieldIndex, byte fieldCount, byte alignment, ushort type, ushort size, ushort secondSize, Type classType)
+        private int AddClass(string name, int fieldIndex, ushort fieldCount, byte alignment, ushort type, ushort size, ushort secondSize, Type classType)
         {
             EbxClass ebxClass = GetClass(classType);
             classTypes.Add(ebxClass);
